backend - functional liked property & frontend - like count in posts without refetching
