+++ components/Modules/Post/UnlikePost.ts
@@ -1,45 +1,32 @@
 import { toast } from 'react-hot-toast';
 
 export type UnlikePostProps = {
-  refetch: () => Promise<void>;
-  localLikeState: string | null;
-  setLocalLikeState: React.Dispatch<React.SetStateAction<string | null>>;
-  localLikeCount: number;
-  setLocalLikeCount: React.Dispatch<React.SetStateAction<number>>;
+  localLikeState: boolean;
+  setLocalLikeState: React.Dispatch<React.SetStateAction<boolean>>;
   id: number;
   unlikeMutation: any;
 };
 
 const UnlikePost = async ({
-  unlikeMutation,
-  refetch,
   localLikeState,
   setLocalLikeState,
-  setLocalLikeCount,
-  localLikeCount,
   id,
+  unlikeMutation,
 }: UnlikePostProps): Promise<void> => {
   // try
   try {
-    // if user liked post before set localLikeCount - 1
-    if (localLikeState !== null) setLocalLikeCount(localLikeCount - 1);
+    if (!localLikeState) return; // if user didn't like post before return
 
-    // set localLikeState to null
-    setLocalLikeState(null);
+    setLocalLikeState(false); // change localLikeState
 
-    // call graphql mutation
-    await unlikeMutation({ variables: { id } });
-
-    // refetch if there are problems with localStates (isn't expected)
-    // await refetch();
+    await unlikeMutation({ variables: { id } }); // call graphql mutation
   } catch (error: any) {
     // throw error if mutation wasn't successful
     toast.error(error.message);
-
     // refetch data
     // (localStates will have wrong value)
     // (possibility that post was deleted)
-    await refetch();
+    // await refetch();
   }
 };
