+++ components/Modules/Post/index.ts
@@ -1,6 +1,5 @@
 export { default as AddToCollectionModal } from './AddToCollectionModal';
 export { default as LikePost } from './LikePost';
 export { default as Post } from './Post';
-export type { default as PostProps } from './PostProps';
 export { default as PostTimeline } from './PostTimeline';
 export { default as UnlikePost } from './UnlikePost';
