+++ lib/graphql/post.graphql
@@ -42,6 +42,7 @@ query personalizedPosts($skip: Int!, $take: Int!) {
       blurhash
       width
       height
+      id
       index
     }
     createdAt
@@ -56,6 +57,8 @@ query post($id: Int!) {
   post(id: $id) {
     id
     liked
+    likeCount
+    commentCount
     createdAt
     postDate
     description
@@ -67,6 +70,7 @@ query post($id: Int!) {
       blurhash
       width
       height
+      id
       index
     }
     author {
@@ -109,10 +113,13 @@ query posts($input: PostsInput) {
       blurhash
       width
       height
+      id
       index
     }
     postDate
     description
+    liked
+    likeCount
     createdAt
     author {
       id
