+++ pages/user/[username]/index.tsx
@@ -50,7 +50,7 @@ const PostsPage: React.FC<{
         },
       },
     },
-  }); 
+  });
 
   return (
     <UserLayout
